+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 * `SelectList` inherits parent width more naturally, and no longer requires an explicit numeric width be set
+* `SelectList` supports `inverted` prop, to expand the selection options upwards rather than downwards
 * `Tag` enables transitions, for consistency with other components
 
 ## v1.2.1